+++ app/Http/Controllers/Resource/DispatcherResource.php
@@ -39,6 +39,7 @@ class DispatcherResource extends Controller
      */
     public function store(Request $request)
     {
+        
         $this->validate($request, [
             'name' => 'required|max:255',
             'mobile' => 'digits_between:6,13',
@@ -98,6 +99,11 @@ class DispatcherResource extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(Setting::get('demo_mode', 0) == 1) {
+            return back()->with('flash_error', 'Disabled for demo purposes! Please contact us at info@appoets.com');
+        }
+        
+
         $this->validate($request, [
             'name' => 'required|max:255',
             'mobile' => 'digits_between:6,13',
@@ -126,6 +132,11 @@ class DispatcherResource extends Controller
      */
     public function destroy($id)
     {
+        if(Setting::get('demo_mode', 0) == 1) {
+            return back()->with('flash_error', 'Disabled for demo purposes! Please contact us at info@appoets.com');
+        }
+        
+
         try {
             Dispatcher::find($id)->delete();
             return back()->with('message', 'Dispatcher deleted successfully');
