+++ app/Http/Controllers/Resource/DocumentResource.php
@@ -39,6 +39,10 @@ class DocumentResource extends Controller
      */
     public function store(Request $request)
     {
+        if(Setting::get('demo_mode', 0) == 1) {
+            return back()->with('flash_error', 'Disabled for demo purposes! Please contact us at info@appoets.com');
+        }
+        
         $this->validate($request, [
             'name' => 'required|max:255',
             'type' => 'required|in:VEHICLE,DRIVER',
@@ -122,6 +126,9 @@ class DocumentResource extends Controller
      */
     public function destroy($id)
     {
+         if(Setting::get('demo_mode', 0) == 1) {
+            return back()->with('flash_error', 'Disabled for demo purposes! Please contact us at info@appoets.com');
+        }
         try {
             Document::find($id)->delete();
             return back()->with('message', 'Document deleted successfully');
