+++ app/Http/Controllers/Resource/UserResource.php
@@ -149,7 +149,12 @@ class UserResource extends Controller
      */
     public function destroy($id)
     {
+         if(Setting::get('demo_mode', 0) == 1) {
+            return back()->with('flash_error', 'Disabled for demo purposes! Please contact us at info@appoets.com');
+        }
+        
         try {
+
             User::find($id)->delete();
             return back()->with('message', 'User deleted successfully');
         }
